+++ euporie/convert/__init__.py
@@ -1,14 +1,25 @@
 """Sub-module concerned with the conversion of data formats."""
 
-from euporie.convert.formats import base64, html, latex, markdown, pil, png, rich, svg
+from euporie.convert.formats import (
+    ansi,
+    base64,
+    jpeg,
+    markdown,
+    pdf,
+    pil,
+    png,
+    rich,
+    sixel,
+)
 
 __all__ = [
+    "ansi",
     "base64",
-    "html",
-    "latex",
+    "jpeg",
     "markdown",
+    "pdf",
     "pil",
     "png",
     "rich",
-    "svg",
+    "sixel",
 ]