+++ euporie/convert/formats/__init__.py
@@ -1,4 +1,4 @@
 """Contains various data format conversion functions.
 
-They are grouped into sub-modules based on starting format.
+They are grouped into sub-modules based on output format.
 """